re-organized employees folder to separate views from controllers
