+++ src/components/employee/views/details.js
@@ -1,5 +1,5 @@
 import React from "react"
-import { modal } from "../../modules/modal/helpers"
+import { modal } from "../../../modules/modal/helpers"
 
 const EmployeeDetails = props => {
     return (
