+++ src/components/employee/views/detailsEditWrapper.js
@@ -1,12 +1,12 @@
 import React from "react"
-import { modal } from "../../modules/modal/helpers"
+import { modal } from "../../../modules/modal/helpers"
 import EmployeeDetails from "./details"
 import EmployeeEditForm from "./editForm"
 import Switch from '@material-ui/core/Switch';
 import FormGroup from '@material-ui/core/FormGroup';
 import FormControlLabel from '@material-ui/core/FormControlLabel';
 import FormControl from '@material-ui/core/FormControl';
-import SuccessSnackbar from "../modal/snackbar"
+import SuccessSnackbar from "../../modal/snackbar"
 
 const DetailsEditWrapper = props => {
     return (
